Meetifyr/pyproject.toml
[tool.poetry]
name = "meetifyr"
version = "0.1.0"
description = ""
authors = ["park2do <ejfhr7777@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
fastapi = "^0.116.0"
orjson = "3.10.12"
uvicorn = "^0.35.0"
sqids = "^0.5.2"
tortoise-orm = "^0.25.1"
cryptography = "44.0.0"
aerich = "0.8.1"
tomlkit = "0.13.2"
pydantic-settings = "2.7.1"
aiomysql = "^0.2.0"
httpx = "0.28.1"

[tool.poetry.group.dev.dependencies]
black = "24.10.0"
ruff = "0.8.2"
mypy = "1.13.0"
pytest = "8.3.4"
coverage = "7.6.9"
pytest-asyncio = "^1.0.0"

[tool.black]
line-length = 50
target-version = ["py311"]
skip-string-normalization = true

[tool.ruff.lint]
select = ["E4", "E7", "E9", "F"]

[tool.mypy]
plugins = ["pydantic.mypy"]
ignore_missing_imports = true
python_version = 3.13
strict = true

[tool.coverage.run]
omit = ["*/test_*.py"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"

[tool.aerich]
tortoise_orm = "app.configs.tortoise_config.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
